+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Version history
 
+## Version 1.1.1 (2026-05-21)
+
+- Added coastlines option for pcolormesh plots.
+- Added default widget values option.
+
+## Version 1.1.0 (2026-05-12)
+
+- Added dpi + share_x|y options.
+
 ## Version 1.0.1 (2025-10-22)
 
 - Fixed issue in get_data() to be able to plot dask arrays.